// src/types/Interfaces-wp/ContactPage.ts
export interface ContactPageData {
  contact_one_location: string;
  contact_one_name: string;
  contact_one_email: string;
  contact_one_number: string;
  contact_two_location: string;
  contact_two_name: string;
  contact_two_email: string;
  contact_two_number: string;
  contact_three_location: string;
  contact_three_name: string;
  contact_three_email: string;
  contact_three_number: string;
  contact_four_location: string;
  contact_four_name: string;
  contact_four_email: string;
  contact_four_number: string;
  contact_five_location: string;
  contact_five_name: string;
  contact_five_email: string;
  contact_five_number: string;
}